SLM-13 Refactor
