+++ src/main/kotlin/uk/gov/justice/digital/hmpps/sendlegalmailtoprisonsapi/magiclink/MagicLinkEmailSender.kt
@@ -3,7 +3,6 @@ package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.magiclink
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.config.MagicLinkConfig
 
 @Service
 class MagicLinkEmailSender(
