+++ src/main/kotlin/uk/gov/justice/digital/hmpps/sendlegalmailtoprisonsapi/security/JwtConfig.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.config
+package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.security
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
